feat: move ReplicateUserFromTable logic to useReplicateUserFromTable. Use new User entity.
