+++ src/webapp/components/replicate/ReplicateUserFromTable.tsx
@@ -1,104 +1,31 @@
-import React, { useCallback, useEffect, useMemo } from "react";
+import React from "react";
 import i18n from "../../../locales";
 import { Id } from "../../../domain/entities/Ref";
-import { useAppContext } from "../../contexts/app-context";
 import { ImportTable, Columns } from "../import-export/ImportTable";
-import { UserProps, defaultUserProps } from "../../../domain/entities/UserProps";
-import { useLoading, useSnackbar } from "@eyeseetea/d2-ui-components";
-import { generateUid } from "../../../utils/uid";
+import { useReplicateUserFromTable } from "../../hooks/useReplicateUserFromTable";
 
 interface ReplicateUserFromTableProps {
     userToReplicateId: Id;
     onRequestClose: () => void;
 }
 
+const columns: Columns[] = [
+    "username",
+    "password",
+    "firstName",
+    "surname",
+    "email",
+    "userRoles",
+    "userGroups",
+    "dataViewOrganisationUnits",
+    "organisationUnits",
+];
+
 export const ReplicateUserFromTable: React.FC<ReplicateUserFromTableProps> = props => {
-    const { compositionRoot } = useAppContext();
     const { userToReplicateId, onRequestClose } = props;
-
-    const [userToReplicate, setUserToReplicate] = React.useState<UserProps>(defaultUserProps);
-    const [isMounted, setIsMounted] = React.useState(false);
-
-    const loading = useLoading();
-    const snackbar = useSnackbar();
-
-    useEffect(() => {
-        const handleUsersError = (message: string) => {
-            snackbar.error(i18n.t(message));
-            onRequestClose();
-        };
-
-        compositionRoot.users.get([userToReplicateId]).run(
-            ([user]) => {
-                if (!user) {
-                    handleUsersError(`Unable to load user: ${userToReplicateId}`);
-                } else {
-                    setUserToReplicate(user);
-                    setIsMounted(true);
-                }
-            },
-            error => {
-                handleUsersError(`Error loading user (${userToReplicateId}): ${error}`);
-            }
-        );
-    }, [compositionRoot.users, onRequestClose, snackbar, userToReplicateId]);
-
-    const replicateTitle = useMemo(() => {
-        console.debug("useMemo replicateTitle");
-        return i18n.t("Replicate {{user}}", {
-            user: userToReplicate ? `${userToReplicate.name} (${userToReplicate.username})` : "",
-        });
-    }, [userToReplicate]);
-
-    const columns: Columns[] = [
-        "username",
-        "password",
-        "firstName",
-        "surname",
-        "email",
-        "userRoles",
-        "userGroups",
-        "dataViewOrganisationUnits",
-        "organisationUnits",
-    ];
-
-    const replicateUsers = useCallback(
-        async ({ users }: { users: UserProps[] }) => {
-            loading.show(true, i18n.t("Replicating users"));
-
-            const newUsers: UserProps[] = users.map(tableUser => {
-                return {
-                    ...userToReplicate,
-                    id: generateUid(),
-                    username: tableUser.username,
-                    password: tableUser.password,
-                    firstName: tableUser.firstName,
-                    surname: tableUser.surname,
-                    email: tableUser.email,
-                    userGroups: tableUser.userGroups,
-                    userRoles: tableUser.userRoles,
-                    dataViewOrganisationUnits: tableUser.dataViewOrganisationUnits,
-                    organisationUnits: tableUser.organisationUnits,
-                    externalAuth: false,
-                    twoFactorEnabled: false,
-                    openId: "",
-                    ldapId: "",
-                };
-            });
-
-            return compositionRoot.users.import({ users: newUsers }).run(
-                () => {
-                    loading.hide();
-                    onRequestClose();
-                    snackbar.success(i18n.t("Users replicated successfully"));
-                },
-                error => {
-                    loading.hide();
-                    snackbar.error(error);
-                }
-            );
-        },
-        [loading, compositionRoot.users, userToReplicate, onRequestClose, snackbar]
+    const { replicateTitle, replicateUsers, userToReplicate, isMounted } = useReplicateUserFromTable(
+        userToReplicateId,
+        onRequestClose
     );
 
     if (!isMounted) {
